+++ src/com/chess/engine/board/BoardUtils.java
@@ -1,6 +1,6 @@
 package com.chess.engine.board;
 
-import com.google.common.collect.ImmutableList;
+import java.util.Map;
 
 public class BoardUtils {
 
@@ -9,8 +9,17 @@ public class BoardUtils {
     public static final boolean[] SEVENTH_COLUMN = initColumn(6);
     public static final boolean[] EIGHT_COLUMN = initColumn(7);
 
-    public static final boolean[] SECOND_ROW = initRow(8);
-    public static final boolean[] SEVENTH_ROW = initRow(48);
+    public static final boolean[] EIGHTH_RANK = initRow(0);
+    public static final boolean[] SEVENTH_RANK = initRow(8);
+    public static final boolean[] SIXTH_RANK = initRow(16);
+    public static final boolean[] FIFTH_RANK = initRow(24);
+    public static final boolean[] FOURTH_RANK = initRow(32);
+    public static final boolean[] THIRD_RANK = initRow(40);
+    public static final boolean[] SECOND_RANK = initRow(48);
+    public static final boolean[] FIRST_RANK = initRow(56);
+
+    public static final String[] ALGEBREIC_NOTATION = initializeAlgebreicNotation();
+    public static final Map<String, Integer> POSITION_TO_COORDINATE = initializePositionToCoordinateMap();
 
     public static final int NUM_TILES = 64;
     public static final int NUM_TILES_PER_ROW = 8;
@@ -46,5 +55,14 @@ public class BoardUtils {
      */
     public static boolean isValidTileCoordinate(final int coordinate) {
         return coordinate >= 0 && coordinate < NUM_TILES;
+    }
+
+    public static int getPositionAtCoordinate(final String position) {
+        return POSITION_TO_COORDINATE.get(position);
+    }
+
+    public static int getPositionAtCoordinate(final int coordinate) {
+        return ALGEBREIC_NOTATION[coordinate];
+
     }
 }