+++ src/com/chess/engine/player/BlackPlayer.java
@@ -67,7 +67,11 @@ public class BlackPlayer extends Player {
                     !this.board.getTile(3).isTileOccupied()) {
                 final Tile rookTile = this.board.getTile(0);
 
-                if (rookTile.isTileOccupied() && rookTile.getPiece().isFirstMove()) {
+                if (rookTile.isTileOccupied() &&
+                        rookTile.getPiece().isFirstMove() &&
+                        Player.calculateAttacksOnTile(2, opponentsLegals).isEmpty() &&
+                        Player.calculateAttacksOnTile(3, opponentsLegals).isEmpty() &&
+                        rookTile.getPiece().getPieceType().isRook()) {
                     kingCastles.add(new QueenSideCastleMove(this.board,
                             this.playerKing,
                             2,